squelch output from shell include (config)
It's bad form to have output from an include like this. If it is
critical that the user be made aware of the heap size (I'm not sure
it is), then it should either be logged from the application or
echoed from bin/cassandra (though the latter is still a bad idea).

Patch by eevans

git-svn-id: https://svn.apache.org/repos/asf/cassandra/trunk@1003202 13f79535-47bb-0310-9956-ffa450edef68