Faster streaming histograms
In ttl-heavy use cases (especially tables with default time to live set), the
streaming histograms calculated during compaction and streaming are very inefficient.

This patch addresses that in two ways:
1) It creates a system property -Dcassandra.streaminghistogram.roundseconds=60,
and rounds all histograms to the next highest multiple of that value, and
2) Rather than maintaining a histogram of 100 bins that have to be merged
on every new value, we keep a temporary spool of 100000 bins, and merge
down to the 100 bin final histogram only when the temporary spool overflows.

Patch by Jeff Jirsa; Reviewed by Nate McCall for CASSANDRA-13038